+++ app/src/views/domain/DomainInfo.vue
@@ -1,9 +1,6 @@
 <template>
-  <div class="domain-info">
-    <b-card>
-      <template v-slot:header>
-        <h2><icon iname="globe" /> {{ name }}</h2>
-      </template>
+  <view-base :queries="queries" skeleton="card-list-skeleton">
+    <card :title="name" icon="globe">
       <!-- VISIT -->
       <p>{{ $t('domain_visit_url', { url: 'https://' + name }) }}</p>
       <b-button variant="success" :href="'https://' + name" target="_blank">
@@ -13,16 +10,12 @@
 
       <!-- DEFAULT DOMAIN -->
       <p>{{ $t('domain_default_desc') }}</p>
-      <template v-if="isMainDomain">
-        <p class="alert alert-info">
-          <icon iname="star" /> {{ $t('domain_default_longdesc') }}
-        </p>
-      </template>
-      <template v-else>
-        <b-button variant="info" v-b-modal.default-domain-modal>
-          <icon iname="star" /> {{ $t('set_default') }}
-        </b-button>
-      </template>
+      <p v-if="isMainDomain" class="alert alert-info">
+        <icon iname="star" /> {{ $t('domain_default_longdesc') }}
+      </p>
+      <b-button v-else variant="info" @click="setAsDefaultDomain">
+        <icon iname="star" /> {{ $t('set_default') }}
+      </b-button>
       <hr>
 
       <!-- DNS CONFIG -->
@@ -41,53 +34,50 @@
 
       <!-- DELETE -->
       <p>{{ $t('domain_delete_longdesc') }}</p>
-      <b-button variant="danger" v-b-modal.delete-modal>
+      <p
+        v-if="isMainDomain" class="alert alert-danger"
+        v-html="$t('domain_delete_forbidden_desc', { domain: name })"
+      />
+      <b-button v-else variant="danger" @click="deleteDomain">
         <icon iname="trash-o" /> {{ $t('delete') }}
       </b-button>
-    </b-card>
-
-    <!-- DEFAULT DOMAIN MODAL -->
-    <b-modal
-      id="default-domain-modal" centered
-      body-bg-variant="danger" body-text-variant="light"
-      @ok="setAsDefaultDomain" hide-header
-    >
-      {{ $t('confirm_change_maindomain') }}
-    </b-modal>
-
-    <!-- DELETE MODAL -->
-    <b-modal
-      id="delete-modal" centered
-      body-bg-variant="danger" body-text-variant="light"
-      @ok="deleteDomain" hide-header
-    >
-      {{ $t('confirm_delete', { name }) }}
-    </b-modal>
-  </div>
+    </card>
+  </view-base>
 </template>
 
 <script>
+import { mapGetters } from 'vuex'
 
 export default {
   name: 'DomainInfo',
+
   props: {
     name: {
       type: String,
       required: true
     }
   },
+
+  data () {
+    return {
+      queries: [{ uri: 'domains/main', storeKey: 'main_domain' }]
+    }
+  },
+
   computed: {
-    mainDomain () {
-      return this.$store.state.data.main_domain
-    },
+    ...mapGetters(['mainDomain']),
 
     isMainDomain () {
       if (!this.mainDomain) return
       return this.name === this.mainDomain
     }
   },
+
   methods: {
-    deleteDomain () {
+    async deleteDomain () {
+      const confirmed = await this.$askConfirmation(this.$i18n.t('confirm_delete', { name: this.name }))
+      if (!confirmed) return
+
       this.$store.dispatch('DELETE',
         { uri: 'domains', param: this.name }
       ).then(() => {
@@ -95,22 +85,17 @@ export default {
       })
     },
 
-    setAsDefaultDomain () {
+    async setAsDefaultDomain () {
+      const confirmed = await this.$askConfirmation(this.$i18n.t('confirm_change_maindomain'))
+      if (!confirmed) return
+
       this.$store.dispatch('PUT',
         { uri: 'domains/main', data: { new_main_domain: this.name }, storeKey: 'main_domain' }
       ).then(() => {
-        // FIXME have to commit here since the response's is empty
+        // Have to commit by hand here since the response is empty
         this.$store.commit('UPDATE_MAIN_DOMAIN', this.name)
       })
     }
-  },
-  created () {
-    this.$store.dispatch('FETCH',
-      { uri: 'domains/main', storeKey: 'main_domain' }
-    )
   }
 }
 </script>
-
-<style lang="scss" scoped>
-</style>
